Add Project to preprocess files
DocFX treats partial classes as one.
Therefor it process and document them as one class
.
In the case of Asserts, there is A LOT of methods,
and it will be dificult to get an overview.

The source code has split them by filename.
I want to be able split them in the same way in
the documentation.